BHBC-1976: Make findById and getSummaryDetails methods in summary-repository; Repurpsoe summary-validation-repo accordingly.
